// src/types/supabase.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: '13.0.5';
  };
  public: {
    Tables: {
// ...
      party_posts: {
        Row: {
          created_at: string | null;
          creator_id: string | null;
          description: string | null;
          game_title: string | null;
          id: number;
          max_members: number | null;
          platform: string | null;
          required_tier: string | null;
          status: string | null;
          title: string | null;
        };
        Insert: {
          created_at?: string | null;
          creator_id?: string | null;
          description?: string | null;
          game_title?: string | null;
          id?: number;
          max_members?: number | null;
          platform?: string | null;
          required_tier?: string | null;
          status?: string | null;
          title?: string | null;
        };
        Update: {
          created_at?: string | null;
          creator_id?: string | null;
          description?: string | null;
          game_title?: string | null;
          id?: number;
          max_members?: number | null;
          platform?: string | null;
          required_tier?: string | null;
          status?: string | null;
          title?: string | null;
        };
        Relationships: [];
      };
// ...
    };
    Views: {
      [_ in never]: never;
    };
    Functions: {
      [_ in never]: never;
    };
    Enums: {
      [_ in never]: never;
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
};